+++ project/Versions.scala
@@ -18,7 +18,7 @@ object Versions {
   val esSpark212 = "8.6.0"
   val scopt = "4.1.0"
   val gcsConnector = "hadoop3-2.2.10"
-  val bigquery = "2.20.1"
+  val bigquery = "2.20.2"
   val hadoop = "3.3.4"
   val sparkBigqueryWithDependencies = "0.28.0"
   val bigqueryConnector = "hadoop3-1.2.0"
@@ -29,6 +29,6 @@ object Versions {
   val confluentVersion = "7.3.1"
   val testContainers = "0.40.12"
   val jsqlparser = "4.5"
-  val gcpDataCatalog = "1.13.0"
+  val gcpDataCatalog = "1.14.0"
   val jinja = "2.6.0"
 }