+++ src/main/scala/ai/starlake/job/sink/bigquery/BigQueryJobBase.scala
@@ -16,7 +16,7 @@ import com.google.cloud.datacatalog.v1.{
   PolicyTagManagerClient,
   PolicyTagName
 }
-import com.google.cloud.{Identity, Policy, Role}
+import com.google.cloud.{Identity, Policy, Role, ServiceOptions}
 import com.google.iam.v1.{Binding, Policy => IAMPolicy, SetIamPolicyRequest}
 import com.google.protobuf.FieldMask
 import com.typesafe.scalalogging.StrictLogging
@@ -28,7 +28,9 @@ import scala.util.{Failure, Success, Try}
 
 trait BigQueryJobBase extends StrictLogging {
   def cliConfig: BigQueryLoadConfig
-  def projectId: String
+
+  def projectId: String =
+    cliConfig.gcpProjectId.getOrElse(ServiceOptions.getDefaultProjectId())
 
   // Lazy otherwise tests fail since there is no GCP credentials in test mode
   lazy val policyTagClient = PolicyTagManagerClient.create()
@@ -67,7 +69,10 @@ trait BigQueryJobBase extends StrictLogging {
   private def applyRLS(forceApply: Boolean)(implicit settings: Settings): Try[Unit] = {
     Try {
       if (forceApply || settings.comet.accessPolicies.apply) {
-        val tableId = TableId.of(cliConfig.outputDataset, cliConfig.outputTable)
+        val tableId = BigQueryJobBase.extractProjectDatasetAndTable(
+          cliConfig.outputDataset,
+          cliConfig.outputTable
+        )
         applyACL(tableId, cliConfig.acl)
         prepareRLS().foreach { rlsStatement =>
           logger.info(s"Applying row level security $rlsStatement")
@@ -168,7 +173,11 @@ trait BigQueryJobBase extends StrictLogging {
               val (location, projectId, taxonomy, taxonomyRef) =
                 getTaxonomy(policyTagClient)
               val policyTagIds = mutable.Map.empty[String, String]
-              val tableId = TableId.of(cliConfig.outputDataset, cliConfig.outputTable)
+              val tableId =
+                BigQueryJobBase.extractProjectDatasetAndTable(
+                  cliConfig.outputDataset,
+                  cliConfig.outputTable
+                )
               val table: Table = bigquery().getTable(tableId)
               val tableDefinition = table.getDefinition().asInstanceOf[StandardTableDefinition]
               val bqSchema = tableDefinition.getSchema()
@@ -278,18 +287,11 @@ trait BigQueryJobBase extends StrictLogging {
     rlsDeleteStatement ++ rlsCreateStatements
   }
 
-  val tableId: TableId = extractProjectDatasetAndTable(
+  val datasetId: DatasetId = BigQueryJobBase.extractProjectDataset(cliConfig.outputDataset)
+
+  val tableId: TableId = BigQueryJobBase.extractProjectDatasetAndTable(
     cliConfig.outputDataset + "." + cliConfig.outputTable
   )
-
-  def datasetId: DatasetId = {
-    scala.Option(tableId.getProject) match {
-      case None =>
-        DatasetId.of(projectId, cliConfig.outputDataset)
-      case Some(project) =>
-        DatasetId.of(project, tableId.getDataset)
-    }
-  }
 
   val bqTable = s"${cliConfig.outputDataset}.${cliConfig.outputTable}"
 
@@ -297,7 +299,7 @@ trait BigQueryJobBase extends StrictLogging {
     val existingDataset = scala.Option(bigquery().getDataset(datasetId))
     val dataset = existingDataset.getOrElse {
       val datasetInfo = DatasetInfo
-        .newBuilder(extractProjectDataset(cliConfig.outputDataset))
+        .newBuilder(BigQueryJobBase.extractProjectDataset(cliConfig.outputDataset))
         .setLocation(cliConfig.getLocation())
         .build
       bigquery().create(datasetInfo)
@@ -378,6 +380,19 @@ trait BigQueryJobBase extends StrictLogging {
     }
   }
 
+}
+
+object BigQueryJobBase {
+
+  def extractProjectDatasetAndTable(datasetId: String, tableId: String): TableId = {
+    BigQueryJobBase.extractProjectDatasetAndTable(datasetId + "." + tableId)
+  }
+
+  /** @param resourceId
+    *   resourceId is a an id of one of the following pattern: <projectId>:<datasetId>.<tableId> or
+    *   <datasetId>.<tableId>
+    * @return
+    */
   def extractProjectDatasetAndTable(value: String): TableId = {
     def extractDatasetAndTable(str: String): (String, String) = {
       val sepIndex = str.indexOf('.')
@@ -411,4 +426,5 @@ trait BigQueryJobBase extends StrictLogging {
       .map(project => DatasetId.of(project, dataset))
       .getOrElse(DatasetId.of(dataset))
   }
+
 }